rewrote the mechanism to write out to the response object.
encoding, prefix and suffix attributes can now be set for any
and all macros (not sure if this will create problems)

Macros handled by objects now check if a function called
handler_macro is defined. If so, the function is called.
Otherwise, the macro evaluates to a property of the object
with the handler name.