Changed remove() method to remove node itself if no arguments are supplied.
Added getNodeProperty() method which will only return properties of the node but not the
underlying prototype, thus avoiding to return functions declared for the prototype. This is used
for the HopObjects get() function.

Added getDefaultValue() to return a string representation for HopObjects. Previously, concatenating
HopObjects to Strings threw an error.