addPart() now takes just one part at a time, but accepts the file name to use for the part as second argument.
Also, it accepts instances of helma.util.MimePart, which are used
in HTTP file uploads and teh global getURL() function.

setText() now plays together nicely with addPart(). If both
plain text and MIME parts are set for a message, the
text will be added as plain text at the beginning of the message.